Move (copy|paste)_files() to ui mainwindow as they are only used there
